Consolidate FormatStringIDLegacy, FormatStringToUpper into Formatting
