+++ src/openrct2-ui/scripting/CustomListView.cpp
@@ -18,7 +18,6 @@
 #    include <openrct2/Context.h>
 #    include <openrct2/localisation/Formatter.h>
 #    include <openrct2/localisation/Formatting.h>
-#    include <openrct2/localisation/Localisation.h>
 #    include <openrct2/util/Util.h>
 
 using namespace OpenRCT2::Scripting;
