+++ src/openrct2-ui/windows/Tooltip.cpp
@@ -16,7 +16,6 @@
 #include <openrct2/drawing/Drawing.h>
 #include <openrct2/localisation/Formatter.h>
 #include <openrct2/localisation/Formatting.h>
-#include <openrct2/localisation/Localisation.h>
 
 namespace OpenRCT2::Ui::Windows
 {
