+++ src/openrct2/config/Config.cpp
@@ -22,8 +22,8 @@
 #include "../drawing/IDrawingEngine.h"
 #include "../interface/Window.h"
 #include "../localisation/Currency.h"
+#include "../localisation/Formatting.h"
 #include "../localisation/Language.h"
-#include "../localisation/Localisation.h"
 #include "../localisation/StringIds.h"
 #include "../network/network.h"
 #include "../paint/VirtualFloor.h"
