+++ src/openrct2/network/NetworkBase.cpp
@@ -79,7 +79,6 @@ static constexpr uint32_t MaxPacketsPerUpdate = 100;
 #    include "../interface/Chat.h"
 #    include "../interface/Window.h"
 #    include "../localisation/Localisation.Date.h"
-#    include "../localisation/Localisation.h"
 #    include "../object/ObjectManager.h"
 #    include "../object/ObjectRepository.h"
 #    include "../scenario/Scenario.h"
