+++ src/openrct2/network/NetworkConnection.cpp
@@ -13,7 +13,6 @@
 
 #    include "../core/String.hpp"
 #    include "../localisation/Formatting.h"
-#    include "../localisation/Localisation.h"
 #    include "../platform/Platform.h"
 #    include "Socket.h"
 #    include "network.h"
