+++ src/openrct2/network/NetworkPlayer.cpp
@@ -12,7 +12,6 @@
 #    include "NetworkPlayer.h"
 
 #    include "../interface/Window.h"
-#    include "../localisation/Localisation.h"
 #    include "NetworkPacket.h"
 
 void NetworkPlayer::SetName(std::string_view name)
