+++ src/openrct2/paint/tile_element/Paint.LargeScenery.cpp
@@ -17,7 +17,6 @@
 #include "../../core/UTF8.h"
 #include "../../interface/Viewport.h"
 #include "../../localisation/Formatting.h"
-#include "../../localisation/Localisation.h"
 #include "../../localisation/StringIds.h"
 #include "../../object/LargeSceneryObject.h"
 #include "../../profiling/Profiling.h"
