+++ src/openrct2/rct1/S4Importer.cpp
@@ -37,8 +37,8 @@
 #include "../entity/Peep.h"
 #include "../entity/Staff.h"
 #include "../interface/Window.h"
+#include "../localisation/Formatting.h"
 #include "../localisation/Localisation.Date.h"
-#include "../localisation/Localisation.h"
 #include "../management/Award.h"
 #include "../management/Finance.h"
 #include "../management/Marketing.h"
