+++ src/openrct2/world/Banner.cpp
@@ -16,9 +16,8 @@
 #include "../core/Memory.hpp"
 #include "../core/String.hpp"
 #include "../interface/Window.h"
-#include "../localisation/FormatCodes.h"
 #include "../localisation/Formatter.h"
-#include "../localisation/Localisation.h"
+#include "../localisation/Formatting.h"
 #include "../management/Finance.h"
 #include "../network/network.h"
 #include "../object/BannerSceneryEntry.h"
